Revamp filesystem monitor implementation
A very big oversight in the previous implementation is that Valen does
not modify the original file ever, it actually deletes it by
clobbering its directory entry to point to a new file with a new inode
number. Thus, we'd normally receive a file deletion event from Valen
and the observer would immediately stop.

This new implementation watches the parent directory so we find out if
the file gets replaced via rename. It's a bit annoying but it works.
It also makes the EidyiaClient side simpler since it doesn't have to
keep track of observers and event handlers itself.