core: General refactoring of async components
This introduces the EidyiaCore class, which is in charge of
orchestrating all asyncio activity in Eidyia and managing three
coroutines: MonitorLoop, DiscordClient, and IrcClient (not yet added
to VCS). Part of the work here is moving the Valen report monitor loop
off from EidyiaDiscordClient and into EidyiaCore so that its operation
can be managed independently of the former, and the services it
provides shared with both EidyiaDiscordClient and EidyiaIrcClient.

Another part of this commit is overhauling the subscriber API since
various resources are now owned by EidyiaCore and shared between
client tasks rather than exclusively owned by EidyiaDiscordClient.

In essence, we are now ready to implement any asyncio frontend to
Eidyia we can think of (although for practical reasons that will only
be IRC for now lol).

There's still some rough edges involving error handling but for now
this is completely functional.